requirements.txt
fitz==0.0.1.dev2
langchain_groq==0.1.9
python-dotenv==1.0.1
streamlit==1.30.0
streamlit==1.38.0
